Extract Dockview logic to DockviewManager
Extracts all Dockview layout initialization, dynamic component
instantiation, layout persistence, DOM resize/overflow detection,
and tab event delegation out of ComposerWorkspace into a dedicated
DockviewManager service.
