+++ shell/src/app/shell/composer-workspace/composer-workspace.ts
@@ -27,19 +27,9 @@ import {
   ElementRef,
   viewChild,
   ViewContainerRef,
-  ComponentRef,
-  Type,
   ChangeDetectorRef,
 } from '@angular/core';
 import {takeUntilDestroyed} from '@angular/core/rxjs-interop';
-import {ChatPanel} from '../../chat/chat-panel/chat-panel';
-import {RawFrame} from '../../preview/raw/raw-frame';
-import {RenderedFrame} from '../../preview/rendered/rendered-frame';
-import {DataModel} from '../../debug/data-model/data-model';
-import {Events} from '../../debug/events/events';
-import {Errors} from '../../debug/errors/errors';
-import {RawMessages} from '../../debug/raw-messages/raw-messages';
-import {MockRules} from '../../debug/mock-rules/mock-rules';
 import {StartupResolution} from '../startup-resolution/startup-resolution';
 import {HostCommunication} from '../host-communication/host-communication';
 import {PreviewBridgeMessageType} from 'a2ui-bridge';
@@ -47,28 +37,15 @@ import {
   AppConfigProvider,
   ThemePreference,
 } from '../../settings/app-config-provider/app-config-provider';
-import {LocalStorageInteractions} from '../../storage/local-storage-interactions/local-storage-interactions';
-import {LocalStorageKey} from '../../storage/models/local-storage-keys';
-import {DockviewComponent} from 'dockview';
+import {ComposerPanelId} from './composer-panel-id';
+import {ComposerDockview} from './composer-dockview.service';
+
+export {ComposerPanelId};
 
 /** Internal interface mapping raw cross-frame workspace telemetry payloads */
 export declare interface WorkspaceMessagePayload {
   action?: unknown;
   validationErrors?: unknown[] | Record<string, unknown> | string | boolean;
-}
-
-/**
- * Defines the possible panel IDs within the Dockview workspace.
- */
-export enum ComposerPanelId {
-  Chat = 'chat',
-  Rendered = 'rendered',
-  Raw = 'raw',
-  DataModel = 'dataModel',
-  Events = 'events',
-  Errors = 'errors',
-  RawMessages = 'rawMessages',
-  MockRules = 'mockRules',
 }
 
 /**
@@ -78,19 +55,18 @@ export enum ComposerPanelId {
 @Component({
   selector: 'a2ui-composer-workspace',
   standalone: true,
+  providers: [ComposerDockview],
   templateUrl: './composer-workspace.ng.html',
   styleUrl: './composer-workspace.scss',
 })
 export class ComposerWorkspace implements OnInit, AfterViewInit {
   private readonly destroyRef = inject(DestroyRef);
-  // Injected to notify zoneless Angular change detection when active panel
-  // signals update.
   private readonly cdr = inject(ChangeDetectorRef);
-  private startupResolution = inject(StartupResolution);
-  private hostComm = inject(HostCommunication);
-  private viewContainerRef = inject(ViewContainerRef);
-  private configProvider = inject(AppConfigProvider);
-  private storage = inject(LocalStorageInteractions);
+  private readonly viewContainerRef = inject(ViewContainerRef);
+  private readonly startupResolution = inject(StartupResolution);
+  private readonly hostComm = inject(HostCommunication);
+  private readonly configProvider = inject(AppConfigProvider);
+  private readonly composerDockview = inject(ComposerDockview);
 
   readonly dockviewRoot = viewChild.required<ElementRef<HTMLElement>>('dockviewRoot');
 
@@ -100,40 +76,18 @@ export class ComposerWorkspace implements OnInit, AfterViewInit {
   unreadErrorsCount = signal(0);
   isDarkTheme = computed(() => this.configProvider.themePreference() === ThemePreference.DARK);
 
-  private readonly isDockviewInitialized = signal(false);
-  private dockviewApi!: DockviewComponent;
-  private componentRefs: ComponentRef<unknown>[] = [];
-
-  private rawMessagesInstance?: RawMessages;
-  private eventsInstance?: Events;
-  private errorsInstance?: Errors;
-
-  private resizeObserver?: ResizeObserver;
-  private animationFrameId?: number;
-
   constructor() {
-    this.destroyRef.onDestroy(() => {
-      if (this.animationFrameId !== undefined) {
-        cancelAnimationFrame(this.animationFrameId);
-      }
-      this.resizeObserver?.disconnect();
-      this.dockviewApi?.dispose();
-      this.componentRefs.forEach(ref => ref.destroy());
-    });
-
     this.hostComm.messageStream$.pipe(takeUntilDestroyed()).subscribe(envelope => {
       if (!envelope) return;
 
       const payload = envelope.payload as WorkspaceMessagePayload | undefined;
-      const eventsPanel = this.dockviewApi?.getGroupPanel(ComposerPanelId.Events);
-      const errorsPanel = this.dockviewApi?.getGroupPanel(ComposerPanelId.Errors);
 
       if (envelope.type === PreviewBridgeMessageType.SEND_TO_SERVER && payload?.action) {
-        if (!eventsPanel?.api.isVisible) {
+        if (!this.composerDockview.isPanelVisible(ComposerPanelId.Events)) {
           this.unreadEventsCount.update(count => count + 1);
         }
       } else if (envelope.type === PreviewBridgeMessageType.CONSOLE_LOG) {
-        if (!errorsPanel?.api.isVisible) {
+        if (!this.composerDockview.isPanelVisible(ComposerPanelId.Errors)) {
           this.unreadErrorsCount.update(count => count + 1);
         }
       } else if (
@@ -147,7 +101,7 @@ export class ComposerWorkspace implements OnInit, AfterViewInit {
             ? Object.keys(validationErrors).length > 0
             : !!validationErrors;
 
-        if (hasErrors && !errorsPanel?.api.isVisible) {
+        if (hasErrors && !this.composerDockview.isPanelVisible(ComposerPanelId.Errors)) {
           this.unreadErrorsCount.update(count => count + 1);
         }
       }
@@ -155,48 +109,26 @@ export class ComposerWorkspace implements OnInit, AfterViewInit {
 
     effect(() => {
       const count = this.unreadEventsCount();
-      const panel = untracked(() => this.dockviewApi?.getGroupPanel(ComposerPanelId.Events));
-      if (panel) {
-        panel.api.setTitle(count > 0 ? `Events (${count})` : 'Events');
-      }
+      this.composerDockview.setPanelTitle(
+        ComposerPanelId.Events,
+        count > 0 ? `Events (${count})` : 'Events',
+      );
     });
 
     effect(() => {
       const count = this.unreadErrorsCount();
-      const panel = untracked(() => this.dockviewApi?.getGroupPanel(ComposerPanelId.Errors));
-      if (panel) {
-        panel.api.setTitle(count > 0 ? `Errors (${count})` : 'Errors');
-      }
-    });
-
-    effect(() => {
-      const show = this.showMockRules();
-      if (!this.isDockviewInitialized()) return;
-
-      const existingPanel = untracked(() =>
-        this.dockviewApi.getGroupPanel(ComposerPanelId.MockRules),
+      this.composerDockview.setPanelTitle(
+        ComposerPanelId.Errors,
+        count > 0 ? `Errors (${count})` : 'Errors',
       );
-      if (show && !existingPanel) {
-        const dataModel = this.dockviewApi.getGroupPanel(ComposerPanelId.DataModel);
-        if (dataModel) {
-          this.dockviewApi.addPanel({
-            id: ComposerPanelId.MockRules,
-            component: ComposerPanelId.MockRules,
-            title: 'Mock Rules',
-            position: {referencePanel: ComposerPanelId.DataModel, direction: 'within'},
-          });
-        }
-      } else if (!show && existingPanel) {
-        existingPanel.api.close();
-      }
     });
 
     effect(() => {
-      const isDark = this.isDarkTheme();
-      const api = untracked(() => this.dockviewApi);
-      if (api) {
-        api.updateOptions({className: isDark ? 'dockview-theme-dark' : 'dockview-theme-light'});
-      }
+      this.composerDockview.toggleMockRules(this.showMockRules());
+    });
+
+    effect(() => {
+      this.composerDockview.updateTheme(this.isDarkTheme());
     });
   }
 
@@ -205,240 +137,25 @@ export class ComposerWorkspace implements OnInit, AfterViewInit {
     this.isExtension.set(isExt);
   }
 
-  ngAfterViewInit() {
-    this.dockviewApi = new DockviewComponent(this.dockviewRoot().nativeElement, {
-      className: this.isDarkTheme() ? 'dockview-theme-dark' : 'dockview-theme-light',
-      defaultRenderer: 'always',
-      createComponent: options => {
-        let type: Type<unknown> | undefined;
-        switch (options.name as ComposerPanelId) {
-          case ComposerPanelId.Chat:
-            type = ChatPanel;
-            break;
-          case ComposerPanelId.Rendered:
-            type = RenderedFrame;
-            break;
-          case ComposerPanelId.Raw:
-            type = RawFrame;
-            break;
-          case ComposerPanelId.DataModel:
-            type = DataModel;
-            break;
-          case ComposerPanelId.Events:
-            type = Events;
-            break;
-          case ComposerPanelId.Errors:
-            type = Errors;
-            break;
-          case ComposerPanelId.RawMessages:
-            type = RawMessages;
-            break;
-          case ComposerPanelId.MockRules:
-            type = MockRules;
-            break;
+  ngAfterViewInit(): void {
+    this.composerDockview.initialize({
+      rootEl: this.dockviewRoot().nativeElement,
+      viewContainerRef: this.viewContainerRef,
+      cdr: this.cdr,
+      destroyRef: this.destroyRef,
+      isDarkTheme: this.isDarkTheme(),
+      showMockRules: this.showMockRules(),
+      onActivePanelChange: panelId => {
+        if (panelId === ComposerPanelId.Events) {
+          untracked(() => this.unreadEventsCount.set(0));
+        } else if (panelId === ComposerPanelId.Errors) {
+          untracked(() => this.unreadErrorsCount.set(0));
         }
-
-        if (!type) {
-          return {element: document.createElement('div'), init: () => {}, dispose: () => {}};
-        }
-
-        const componentRef = this.viewContainerRef.createComponent(type);
-        this.componentRefs.push(componentRef);
-
-        if (type === RawMessages) this.rawMessagesInstance = componentRef.instance as RawMessages;
-        if (type === Events) this.eventsInstance = componentRef.instance as Events;
-        if (type === Errors) this.errorsInstance = componentRef.instance as Errors;
-
-        return {
-          element: componentRef.location.nativeElement,
-          init: params => {
-            componentRef.changeDetectorRef.detectChanges();
-          },
-          dispose: () => {
-            if (componentRef.instance === this.rawMessagesInstance) {
-              this.rawMessagesInstance = undefined;
-            }
-            if (componentRef.instance === this.eventsInstance) {
-              this.eventsInstance = undefined;
-            }
-            if (componentRef.instance === this.errorsInstance) {
-              this.errorsInstance = undefined;
-            }
-            componentRef.destroy();
-            this.componentRefs = this.componentRefs.filter(r => r !== componentRef);
-          },
-        };
       },
-    });
-
-    this.dockviewApi.onDidActivePanelChange(event => {
-      const panel = event.panel;
-      if (panel?.id === ComposerPanelId.Events) {
-        untracked(() => this.unreadEventsCount.set(0));
-      } else if (panel?.id === ComposerPanelId.Errors) {
-        untracked(() => this.unreadErrorsCount.set(0));
-      }
-      this.checkTabOverflow();
-      // Notify zoneless Angular change detection when active panel signals update.
-      this.cdr.markForCheck();
-    });
-
-    const savedLayout = this.storage.getItem(LocalStorageKey.DOCKVIEW_LAYOUT);
-    let layoutRestored = false;
-
-    if (savedLayout) {
-      try {
-        this.dockviewApi.fromJSON(JSON.parse(savedLayout));
-        layoutRestored = true;
-      } catch (e) {
-        console.error('Failed to restore dockview layout', e);
-      }
-    }
-
-    if (!layoutRestored) {
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.Chat,
-        component: ComposerPanelId.Chat,
-        title: 'Gemini Assistant',
-      });
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.Rendered,
-        component: ComposerPanelId.Rendered,
-        title: 'Rendered A2UI Preview',
-        position: {direction: 'right', referencePanel: ComposerPanelId.Chat},
-      });
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.Raw,
-        component: ComposerPanelId.Raw,
-        title: 'A2UI JSON Editor',
-        position: {direction: 'right', referencePanel: ComposerPanelId.Rendered},
-      });
-
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.DataModel,
-        component: ComposerPanelId.DataModel,
-        title: 'Data Model',
-        position: {direction: 'below', referencePanel: ComposerPanelId.Rendered},
-      });
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.Events,
-        component: ComposerPanelId.Events,
-        title: 'Events',
-        position: {direction: 'within', referencePanel: ComposerPanelId.DataModel},
-      });
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.Errors,
-        component: ComposerPanelId.Errors,
-        title: 'Errors',
-        position: {direction: 'within', referencePanel: ComposerPanelId.DataModel},
-      });
-      this.dockviewApi.addPanel({
-        id: ComposerPanelId.RawMessages,
-        component: ComposerPanelId.RawMessages,
-        title: 'Raw Messages',
-        position: {direction: 'within', referencePanel: ComposerPanelId.DataModel},
-      });
-
-      if (this.showMockRules()) {
-        this.dockviewApi.addPanel({
-          id: ComposerPanelId.MockRules,
-          component: ComposerPanelId.MockRules,
-          title: 'Mock Rules',
-          position: {direction: 'within', referencePanel: ComposerPanelId.DataModel},
-        });
-      }
-    }
-
-    let saveTimeout: ReturnType<typeof setTimeout>;
-    this.dockviewApi.onDidLayoutChange(() => {
-      this.checkTabOverflow();
-      clearTimeout(saveTimeout);
-      saveTimeout = setTimeout(() => {
-        this.storage.setItem(
-          LocalStorageKey.DOCKVIEW_LAYOUT,
-          JSON.stringify(this.dockviewApi.toJSON()),
-        );
-      }, 1000);
-    });
-    this.dockviewApi.onDidAddPanel(() => this.checkTabOverflow());
-    this.dockviewApi.onDidRemovePanel(() => this.checkTabOverflow());
-
-    this.resizeObserver = new ResizeObserver(() => this.checkTabOverflow());
-    this.resizeObserver.observe(this.dockviewRoot().nativeElement);
-
-    // Force an initial layout pass. In browsers, ResizeObserver handles this,
-    // but in jsdom tests with mocked observers, it requires an explicit call.
-    const width = this.dockviewRoot().nativeElement.clientWidth || 1000;
-    const height = this.dockviewRoot().nativeElement.clientHeight || 1000;
-    this.dockviewApi.layout(width, height);
-    this.isDockviewInitialized.set(true);
-    this.checkTabOverflow();
-
-    const handleTabInteraction = (event: Event) => {
-      const tabEl = event
-        .composedPath()
-        .find(
-          (node): node is HTMLElement =>
-            node instanceof HTMLElement && node.classList.contains('dv-tab'),
-        );
-      if (!tabEl) return;
-
-      const panels = this.dockviewApi?.panels ?? [];
-      for (const panel of panels) {
-        const panelTabEl = panel.view?.tab?.element;
-        if (panelTabEl && (panelTabEl === tabEl || panelTabEl.contains(tabEl))) {
-          if (panel.api && !panel.api.isActive) {
-            panel.api.setActive();
-          }
-          break;
-        }
-      }
-    };
-
-    const rootEl = this.dockviewRoot().nativeElement;
-    // Register capture-phase pointerdown and click event delegation on
-    // #dockviewRoot to bridge Dockview's internal pointerdown contract and
-    // ensure tab panel activation via panel.api.setActive().
-    rootEl.addEventListener('pointerdown', handleTabInteraction, true);
-    rootEl.addEventListener('click', handleTabInteraction, true);
-
-    // Register automatic event listener cleanup when the component is destroyed
-    // via destroyRef.onDestroy.
-    this.destroyRef.onDestroy(() => {
-      clearTimeout(saveTimeout);
-      rootEl.removeEventListener('pointerdown', handleTabInteraction, true);
-      rootEl.removeEventListener('click', handleTabInteraction, true);
-    });
-  }
-
-  /**
-   * If the tabs don't fit in the available space, then we want the dockview
-   * overflow selector to be displayed. Otherwise, it's redundant, so it
-   * should be hidden. Unfortunately, there doesn't seem to be a "native"
-   * mechanism in dockview to support this.
-   */
-  private checkTabOverflow(): void {
-    if (this.animationFrameId !== undefined) {
-      cancelAnimationFrame(this.animationFrameId);
-    }
-    this.animationFrameId = requestAnimationFrame(() => {
-      this.animationFrameId = undefined;
-      const rootEl = this.dockviewRoot()?.nativeElement;
-      if (!rootEl) return;
-      const tabContainers = rootEl.querySelectorAll<HTMLElement>('.dv-tabs-container');
-      let hasOverflow = false;
-      tabContainers.forEach(container => {
-        if (container.scrollWidth > container.clientWidth + 2) {
-          hasOverflow = true;
-        }
-      });
-      rootEl.classList.toggle('has-tab-overflow', hasOverflow);
     });
   }
 
   clearAllLogs(): void {
-    this.rawMessagesInstance?.clearLogs();
-    this.eventsInstance?.clearLogs();
-    this.errorsInstance?.clearLogs();
+    this.composerDockview.clearAllLogs();
   }
 }
